backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
python-dotenv==1.0.1
ccxt==4.4.40
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
xgboost==2.1.3
scikit-learn==1.5.2
torch==2.5.1
httpx==0.28.1
apscheduler==3.10.4
sentry-sdk[fastapi]==2.19.2
alembic==1.14.0
cryptography==44.0.0
prometheus-client==0.21.1
langfuse==2.59.7
transformers==4.46.3
